+++ tools/runners/run-velocity.js
@@ -1,9 +1,18 @@
 var Console = require('../console/console.js').Console;
 var isopackets = require('../tool-env/isopackets.js');
-
-var phantomjs = require('phantomjs-prebuilt');
 var child_process = require('child_process');
 var _ = require('underscore');
+
+try {
+  var phantomjs = require('phantomjs-prebuilt');
+} catch (e) {
+  throw new Error([
+    "Please install PhantomJS by running the following command:",
+    "",
+    "  meteor npm install -g phantomjs-prebuilt",
+    ""
+  ].join("\n"));
+}
 
 // XXX this could really use a self-test!